Removing Mount, adding Stream support to configuration files, implementing plugin loader, adjusting plugins, adding class loader.
The plugin loader first reads all jar files for their CANARY.INF. It assigns the mount point and later will do dependency checking there. Then it loads at the appropriate times the loading of the plugin by finding the manifest for its main-class.

post-loaded plugins are not postloaded and enabling/disabling is not implemented. Nor is dependency checking. (TODO)