Changes in Backbones, Player and xml
Added a piece of code that will cause a schema update upon initialisation of a backbone.
Added modular sub groups to player.
Changed XML file writing a little which is an attempt to counter
data corruption that may occur while saviong to backend.
Also this needs testing on win8 because of the file lock issue.
It might be back. Cannot test myself, I have no Win8