Changes in Plugin Config, new Command, changes in Player and OfflinePlayer
OfflinePlayer and Player now have methods to remove groups from them.
This will not allow to remove the main group of a player!

Configuration.getPluginConfig(*) now stores plugin configs in config/PLUGINNAME/...
instead of in the root directory.

Changed the LinkedList in Servertastmanager to ArrayList.

Added command to remove a group from a player