Read and create ops.cfg
This is not synchronous with the real ops list in database,
however, a user can utilise this to quickly set up an initial list
of operators without meddling with commands of any kind.

It will also be read everytime the operatorsProvider is reloaded or instantiated.
So new names inside this file will always be recognised.